+++ bika/lims/browser/analysisrequest/reject.py
@@ -22,10 +22,16 @@ from bika.lims.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 
 
-class AnalysisRequestRejectBase(object):
+class AnalysisRequestRejectPdfView(BrowserView):
     """
-    Provides helper methods that ease the work with rejection reasons
+    View that renders the template to be used for the generation of a pdf to
+    be attached in the email for the notification of an Analysis Request
+    rejection action.
     """
+    template = ViewPageTemplateFile("templates/analysisrequest_retract_pdf.pt")
+
+    def __call__(self):
+        return self.template()
 
     def get_rejection_reasons(self, keyword=None):
         """
@@ -49,33 +55,3 @@ class AnalysisRequestRejectBase(object):
             return rejection_reasons.get(keyword, '') and [rejection_reasons.get(keyword, '')] or []
         return rejection_reasons.get(keyword, [])
 
-
-class AnalysisRequestRejectEmailView(BrowserView, AnalysisRequestRejectBase):
-    """
-    View that renders the template to be attached in the body of the email
-    for the notification of an Analysis Request rejection action.
-    """
-
-    template = ViewPageTemplateFile("templates/analysisrequest_retract_mail.pt")
-
-    def __init__(self, context, request):
-        super(AnalysisRequestRejectEmailView, self).__init__(context, request)
-
-    def __call__(self):
-        return self.template()
-
-
-class AnalysisRequestRejectPdfView(BrowserView, AnalysisRequestRejectBase):
-    """
-    View that renders the template to be used for the generation of a pdf to
-    be attached in the email for the notification of an Analysis Request
-    rejection action.
-    """
-
-    template = ViewPageTemplateFile("templates/analysisrequest_retract_pdf.pt")
-
-    def __init__(self, context, request):
-        super(AnalysisRequestRejectPdfView, self).__init__(context, request)
-
-    def __call__(self):
-        return self.template()